+++ src/js/components/perfPages/formFields/index.js
@@ -113,3 +113,37 @@ export const createField = (
   return null;
   // throw new Error(`Unknown field type: ${filterDef.field_type}`);
 };
+
+// This is a gross hack but its a legitimately hard problem: the server
+// controls what fields it sends but the client needs to make it look nice.
+// I need the fourth field (which I know is "Success Percentage") to wrap
+// to the next line.
+//
+// There is no good place to put the responsibility for deciding how to lay
+// things out. In the future we will use a filter UI more like Salesforce's
+// and this hack can go away.
+//
+// Essentially random layout of server fields is why so many enterprise apps
+// look horrible. This small hack is a compromise. It means that the client
+// and server are more tightly bound then they otherwise would be. Adding
+// or removing fields requires a change here.
+const spacerField = {
+  name: 'Blank "field" to space things out',
+  currentValue: '',
+  render: () => <span />,
+};
+
+export const AllFilters = ({ filters }: { filters: Field[] }) => {
+  // Yes...this is gross. I'm sorry.
+  filters.splice(3, 0, spacerField);
+  filters = filters.filter(x => x); // get rid of nulls
+  return (
+    <div key="filterGrid" className="slds-grid slds-wrap slds-gutters">
+      {filters.map(filter => (
+        <div key={filter.name} className="slds-col slds-size_3-of-12">
+          {filter.render()}
+        </div>
+      ))}
+    </div>
+  );
+};
